Initial revision `word_tools.urban`
This function accepts a word and optional result limit argument.
UrbanDictionary is queried using `requests` module, and the response is
then parsed using `BeautifulSoup`. Returns a list of results up to
`limit` results.

Currently, `limit` can properly limit the number of results returned.
But does not properly 'return all' when using the default of `0`.

This can use some more clean up, but it works. Committing here as a
functional checkpoint in case something goes horribly wrong with
cleaning up and fixing `limit`.