`limit` argument now properly functional
When `urban.lookup()` is called with a `limit <= 0`, we return all
results. This is an improvement over the original behavior that would
return zero results when the limit was less than zero. I mean,
technically, that would be correct functioning of the limit. But not
intended behavior.

Also update the test case to print the total number of results following
their listing.