Begin CLI entry-points into `word_tools`
Got some new files for this one, exciting. Create a new sub-module
within `word_tools` to handle all of the CLI action. This is intended to
be a simple entry-point into the module functions. When paired up with a
properly configured `setup.py`, sort of aliases into specific functions
can be created. The alias style may be preferred for now as that
eliminates the need to create a larger-scope, fully-featured CLI
application.

`setup.py` coming up next..