Remove `__main__.py`
With the `setup.py` entry-points, we do not have a need for
`__main__.py` at this moment. Keeping it around might have caused
problems, and likely would have caused me confusion at some point. It
can easily be re-written if need be in the future as it simply called
our `word_tools.cli.main()` function.