Let's try generators!
After some reading up on generators, for probably the third or fourth
time since I started learning Python.. I finally felt I saw an
opportunity, and felt comfortable enough to use one. I have no idea if
this is faster, or more effective or anything. But it does cut back on
one list being created, and it slightly adjusts the program flow.

Now, rather than returning a single list of all results, we create a
generator which returns single results at a time. In essence, this
functions as an iterator over the results, returning one at time. The
calling code determines how many results it should get.

The `limit` argument will be removed from the `word_tools.lookup`
methods in a future commit. They were left in to limit the changes made
to the code at once, though this really is a small commit. Keeping it
byte-sized.