Very rough draft wikipedia lookup
Functional in the sense it works, unfinished in that we don't get the
output we would like. The `lookup.wikipedia()` function follows the same
format as the other two lookups. Accept a word, and optional limit.
Return results. This is slightly different in that we are also going to
return a URL to the result page. In the future, the other lookup
functions may be updated to include that functionality.

Touch up the `__main__` block used for our test case. Just changing up
the format, and updating to work with the tuple now being returned
instead of individual items.