`url_to_soup` bad request test properly handled
Originally, the implemenation of `test_url_to_soup_failure` was not
truly testing `lookup.url_to_soup`. It was more testing the `requests`
module. While attempting to fix up the loop a bit, I realized that
issue and implemented a fix.

Rather than `assert` based on the value within the `HTTPError` exception
itself, we simply ensure that the exception has been thrown. Our
particular case does not seem to require checking the actual exception
text. As a effect of properly testing for this exception, we have
considerably cleaned up the `for` loop. It all works out.