Test `lookup.urban_dictionary`
Think I *finally* got this one figured out. A bit tricky to test for,
because it depends on the content returned by Urban Dictionary, out of
our control. By choosing lesser known words, we may lower our chances of
the results updating and invalidating our test.

Otherwise, this is a series of `for` loops, where the outermost will
iterate into our `words` dictionary. That dictionary stores as keys the
`word` to test, and a list containing a `limit` first item and fragments
of `results` in the remaining list items. inside the loop, we call
`lookup.urban_dictionary`, then enter another `for` loop which grabs
results and checks for each fragment within the results.

All assertions return true to indicate success.