Test `lookup.merriam_webster` and `lookup.wikipedia`
Following the same format as `test_urban_dictionary`, we bring in our
final two `lookup` functions for testing. Took a slight bit more thought
to get the Wikipedia test functional, but it works now.

How complete is this coverage? I haven't the slightest idea, doubt it's
full. But it's enough for now. Our core functionality, and Python sanity
can be tested and all tests pass properly. Sounds like a good time to
see what's up with Travis CI, because why not?