It would help if I actually submit the report..
Everything worked flawlessly, quite a rush to watch the commit just
pushed clone, install and test itself live from the Travis CI jobs
screen. Then it got to the end, and I realized the coverage reports
aren't actually being submitted. Oops.

Add in the `codecov` command so that they are actually submitted on
successful Travis runs. May as well add a comment as well. And a `-` to
the `pytest` command, as that seems to be the usual convention. I think
I need to read into the meaning of that. Not very familiar with `.yml`
syntax, at all.

Let's see if we get coverage now!