Specify linux distro, add test for MacOS
While the current default Linux distribution for Travis *is* Xenial, we
can still specify it anyway. A little extra clarification never hurts
should the default change unexpectedly at some point.

The main feature, in this commit, including a job for Python 3.6.5 on
MacOS 10.13. This should work, at least if the references I used are
correct, and any number of other factors totally outside my thoughts
right now. I haven't used Mac OS in years, so I'm totally unfamiliar
with the setup and use of Python. The implementation of this job was
entirely influenced from a blog post, linked below.

In all, this seems like a good idea to include multiple OS test jobs. I
may not be able to physically test the code myself on various setups.
But if Travis can, that would be really helpful!

  - https://blog.travis-ci.com/2019-08-07-extensive-python-testing-on-travis-ci