Specify `bionic` target for Travis. Update indents.
Update Travis to use the more recent Ubuntu image. Not sure how profound
the difference would be between the minor version changes. Feels to me
like the more up to date would be better, but I could be wrong.

Now that I have also read a bit on the rules of `.yml` formatting, I see
I can cut my indenation in half. There was thought to use 2 spaces per
level of indent. Settled on 4 spaces per indent as it matches the Python
sources.