Stub `test_lookup_limit`
Stub for test of the `limit` argument in our `lookup` functions.
Intended behavior is to recieve a `StopIteration` error if we attempt to
reach into our results further than `limit` specifies. Test that.

I envision this by using a `for` loop to iterate over a list of
function. Each function is called, and the results grabbed some number
of times until we hit a `StopIteration`, assert that, and (hopefully)
pass the test.