Test for bad input types into `stoopid`
When given input types such as lists, integers, or dictionaries,
`stoopid` should fail by throwing a `TypeError` back at us. This test
ensures that is the case.

As of this commit, it seems to be doing it's job. It has highlighted an
error with our code. Rather than refusing to work with lists,
dictionaries, tuples, or other iterables, `stoopid` attempts to work it
out. This results in unexpected, or incorrect behavior. We will need to
revisit `stoopid` and see if we can't knock some sense into it.