Add `stoopid` entry-point
Find a nice, stubby entry-point and attach it to `stoopid`. The `extra`
argument is not properly accepted in this code, but it doesn't work in
the main `stoopid` function either so there's no difference either way
for the time being.

This commit brings in the ability to `wt.stoopid` from the terminal,
rather than firing up a REPL.