Test limit in `lookup` providers
*Finally!* Hitting a point where I can feel comfortable commiting. Took
me a solid 90 minutes of playing with this function, moving around lines
of code, and banging my head on the keyboard. It seems to work now.
Well, mostly.

In the case of a `limit == 0`, we should receive all results from the
provider. At least, the first page of results anyway. I was unable to
get that edge case tested properly, thus it is being omitted for the
time being. Due to the struggles I just had, I am thinking it may be a
candidate for it's own specific test.

Otherwise, the gist of this.. Define a dictionary containing providers
in `lookup` to test, function keys and limit values. Iterate over the
items in that dictionary, first calling `lookup.<provider>()` and then
entering a `while` loop to grab results. Once we grab all results, push
the limits one more time. But make sure we get a `StopIteration` in
return for our antics.