Update test search terms
Spent yet another couple of hours back and forth with this one. Trying
all sorts of various things from `any()` to comprehrensions. Just could
not get this working the way I originally intended. For now, the fix is
to replace the 'python' search terms with 'fraction', as those results
seem less dynamic than Python's.

The original intention is to compare each result to each search term. We
would provide, say 5 sets of fragments per search term. And while we may
only test the first couple results, we will have more freedom for the
results to re-order themselves due to Wikipedia being dynamic. I will
continue looking in to a more complete solution to this. For now, the
workaround is in place and seems to handle as it should.