Add imports to `word_tools` module init
Decided today is a good day to read into the proper use of `__init__.py`
in regards to Python modules. Turns out, this is how you fill in the
module's namespace. And it fixes the import errors I experienced when
initially writing the module.

This will allow users of the module to `import word_tools` without
having to worry about sub-modules or the like.