New import style for `transform` tests
Following suit of the last commit, the `transform` tests now use the
updated import syntax. Simply results in easier to follow code.

Note that I attempted to refactor the `lookup` tests as well. Ran into
problems with `lookup.url_to_soup` not being exported as part of the
`import word_tools` call. I was unable to determine a solution that did
not involve adding `url_to_soup` into `word_tools/__init__.py` to be
part of the main module. Look into it further and determine the next
course of action.