Preparing for PyPi packaging
We cover a decent bit of ground, and progress through this commit. This
covers most of the requirements to create a wheel, and upload it to PyPi
via twine. Using the test instance of PyPi, I notice an error regarding
`bs4` not being found when `pip` attempts to install requirements. I
need to look into that further.

However, installing via PyPi seems to prevent the 'word_tools not found`
ImportError we were previously getting from `pip install .`. That's a
plus.

Otherwise, breaking this commit down:
  - Switch import style, only take what we need
  - Pull README.md into setup()
  - We figured out what `find_packages()` does
  - Specify license, keywords
  - Add more classifiers
  - Typo: Independant -> Independent