Append missing `,`
I had missed a comma at the end of an item in our classifiers list.
Turns out in that case, the strings concatenate themselves rather than
throwing a syntax error. Good to know.

Uploaded, then installed from test instance of PyPi sucessfuly. There
was an error regarding a specific version mismatch with `soupsieve` but
that was able to install manually. The error was simply due to the
instance not having the exact version numbers as the main PyPi site.