Create CHANGELOG.md, include in MANIFEST.in
Now that we've got releases beginning to work their way out, it's a good
time to include a changelog. This will be appended to using the GitHub
Changelog Generator linked in the file itself. I haven't totally figured
out the tool yet, so manual editing will almost definitely be necessary.
But I like it's output, and it makes things a bit simpler.

Let's also include `CHANGELOG.md` on the manifest so it is included in
our release packages.