Add contributing guidelines
This beast has been a long time in the works. Without any real work
until an hour ago. Having gone from 'only typing a quick stub' to 'a
damn fine essay', I think we've accomplished our goal.

This document serves to assist a person who wishes to contribute to this
project but is not sure where to begin. Feedback is welcome and this
document will be expanded upon in time, I'm sure.