Add `push.yml` -- code linting
GitHub Actions caught my attention some time ago. They always felt
complex, and out of my grasp. They still kind of do. This is an attempt
to leverage their power and use them in a project for great good. Having
spent a fair bit of time reading documentation and searching for
solutions. This is what I've come up with.

This action will run on every push event and perform a basic code lint
check. Specific syntax errors or undefined names will cause the check to
faiil. Other formatting errors will be displayed in the log, but not
cause a failure.

I am uncertain if this will block a push from happening upon failure, or
not. My feeling is that it will display in a checks page, similar to
Travis CI, with log output and status.

Let's give this a shot, and see how it works out. Fingers crossed I
didn't break anything too bad ;)