Correct typo, remove explicit version check.
It worked! Now that we have verified our action works, we can verify
it's operation. The main thing that stood out to me -- Python versions.
The original revision of this file directly specified Python 3.6. I'd
like to just use the latest minor version of Python 3. This is more
because it isn't exactly important which minor version we use, as long
as it is >=3.6.

Use this opportunity to also remove the explicit Python version check.
When putting this action together, I anticipated that we might not have
the Python version displayed after setting it up. Turns out the version
number is displayed, so this check is not necessary. Remove it.