Add label sync'ing GitHub Action
There seem to be a wide variety of options available for this feature.
Attempting to keep things simple, `micnncim/action-label-syncer` is the
solution I felt right at this time.

This action will run on each push to this repo, and synchronize the
labels within `.github/labels.yml` to the GitHub interface. Simple, easy
updating and maintenance of your label list. The big draw for me, aside
from more leveraging of GitHub Actions, is that this will make importing
our labels into new projects *much* easier.

Let's see if this works!