Include `periodic-labeler` action
Let's try out the `periodic-labler`. In this configuration, a check is
run every 12 hours. Any pull requests are matched against the patterns
specified in `.github/labeler.yml`, and any matches have the
cooresponding label applied.

This should work on both PRs originating from this repo, and PRs
originating from forked repos. Will need to verify, but I think it will
be just fine.

These rules and patterns will need updating as time goes on, and the
cron specification may be updated. For now, this is a good start and I
think it will help some of the busy work automate itself away.