Issue #2 -- until fixed, allow test failure
Planned fix is pending a refactor of `word_tools.lookup` to use a system
of 'lookup providers' and a lookup factory sort of deal to instantiate
them. This (and other providers tests!) will require some amount of
re-writing to handle the dynamic nature of web search results.

See issue #7 for further details on the changes proposed to `lookup`.