Implement basic factory/provider pattern
This one has been a long time in the making, finally making progress
minus a few delays. Having begun to feel some growing pains in
`word_tools` as I began to expand the test suite, and look into adding
new functionality, I figured there must be a better way. A more
maintainable, and sensisble way. Enter the object factory.

The pattern, as implemented here, provides `LookupProvider` and
`TransformProvider` interfaces which are used as the base for classes
such as `UrbanProvider` or `StoopidProvider`. Our interface provides a
blueprint of *what* each derived provider must be able to do, but does
not define *how*. Put another way, the basic providers define empty
abstract methods which must be overriden by the class derived from the
base provider.

During initialization of the main `word_tools` module, we register each
provider into a dictionary where keys are the provider names, and values
are the provider object themselves. Each provider is implemented as a
singleton, therefore only one instance of it may exist at any given
time. Individual providers may be 'grabbed' with an expression such as:

stoopid = word_tools.transform.get('stoopid')
urban = word_tools.lookup.get('urbandictionary')

Which will look for the derived provider stored in the
`object_factory.providers` dictionary. If a match is found, return an
instance of the class. The returned object may be used according to the
interface definied in `LookupProvider` or `TransformProvider`. Which
brings up another point, this decouples us from knowing which specific
provider we are working with until runtime. Each provider type maintains
it's own interface, and derived providers must implement it. How it
works, does not matter. That it's public interface is guaranteed to
match other providers, does matter.

In essence, this is the beginning of a mostly full re-write of
`word_tools`. Allowing us greater room to grow, and expand over time.
And providing us with more maintainable code, that repeats itself less.
It's been a hell of a lot of work setting this initial state up. Have
been quite a few scrapped beginnings and `git reset`s along the way. I'm
glad to have reached a point where I feel comfortable committing this to
our history.

Remaining work will include adapting our tests, and CLI to utilize the
new system of providers. I am planning as part of the CLI integration,
to do away with our original entry-point system. Further details on that
will be in a yet to be written issue. Issue is inbound shortly though.

Note that for now, I am leaving Wikipedia out of the picture. That
provider will require writing it over from the beginning. Rather than
directly processing the results ourself, we should be using the
`wikipedia` module as it provides a proper API. And will be *much*
easier to work with, and to have functional tests for. Further details
on that are filed under Issue #2.

Whew. One hell of a commit message. If you read this entirely, thank
you. I don't expect most to though. When I come back in another year,
wondering what the hell I was thinking -- this will be my saving grace.