Prefer specific provider over generic provider
Rather than using the `GenericProvider` interface with our
`ObjectFactory`, we should use the proper interface. The abstract
methods should still require being overridden because the concrete
providers all derive from the interface.

The rationale behind this change is to be a bit more explicit in our
code, and therefore readable and (possibly) less likely to run into
errors down the road.