Update `url_to_soup` tests for new module layout
Beginning now some updates to our tests so that they work with the new
module layout. Fairly simple change here. Simply create the base
`LayoutProvider`, and call `url_to_soup`. Because we are not deriving
from this class, we do not need to override the abstract methods -- they
are already implemented as blank functions.

Also, because this function is part of the base provider for all lookup
providers, we do not need to test individual derived classes. They
inherit the method being tested. Thus by testing the parent, we ensure
our children are safe.