Update import style now, reduce code later
By importing the entire `word_tools` module, rather than picking out
specific bits and pieces, we can leverage the setup contained in
`__init__.py`. Setup being the registration of providers, and import of
necessary submodules. This will prevent the need for us to perform those
steps ourselves in each test.

I *think* that is the proper way to do it. That would mean we have
coverage on our module init code by association. Probably a good thing.
If we re-write the init in the specific tests, we introduce code
repetition and lose some coverage. I may be wrong on all of this. It
seems about right though.