Fix final tests in `test_lookup.py`
These are the last few functions to be updated for the new module
layout. This includes `test_{urban_dictionary,merriam_webster,wikipedia}`
providers. For now, wikipedia has been updated, but is still an expected
failure.

Surprisingly, updating all of these tests was very straight-forward. The
core change, to the user of this module, is that we utilize objects
rather than a functional design. The API was kept nearly the same so
these updates went very smooth.