Docstring overhaul. Remove argument from provider.
We needed docstrings. Badly. Wrote some up. Probably not the best. But
they are better than the `None` we had before. The plan is the
*eventually* write a more proper API documentation.

Remove the `extra` argument from `TransformProvider.convert` abstract
method. This is not needed as other `TransformProvider`s may not want to
use that argument. If `StoopidProvider` wants to accept that argument,
it's on them. We won't stop them. But we won't require anyone else to
accept it, and we won't either.