Implement `WikiProvider`
Add the new implementation of `WikiProvider` -- powered by
`pymediawiki`! This module seems much nicer than the home-grown solution
we previously used. In `__init__`, we set up a global `MediaWiki` object
for this provider to use. And over in `lookup` we assemble a string
featuring the title, url and trimmed summary before yielding each
result.