Hello Disambiguation, my old friend.
Ahh, just couldn't get away from the `DisambiguationError` exception,
could i? Turns out this is still a thing in this module as well.
However, I have decided a proper way to handle the error this time
around. If encountered, display a list of possible terms. The core issue
seems like it may just be something related to Wikipedia itself. Need
further investigation to say for sure.

Also, set a default limit of three results when no argument is
specified. `pymediawiki` handles the `results` argument quite literally.
If `0`, or `None` are passed in you get just that.. No results. But you
do get errors, if that's your thing. For the time being, circumvent by
using a valid limit as the default.