Fix `test_wikipedia` -- it works!
Much thanks to the `pymediawiki` module for making this test *much*
simpler to re-write. I'm enjoying this module, for sure. Rather than the
original, convoluted strategy of checking for result fragments in each
result, and hoping the results never update. We instead retrieve results
and simply check that our original search term shows up in the results.
This should work just fine -- but if anyone knows better, please inform.

This new strategy should no longer be effected by the slightest breeze
in the wind, while still maintaining that we have recieved at least
semi-valid results. Using mock data would be the true way to test the
results. I'm not sure I'm ready for that yet..