Interactive staging misadventure. Pre-merge lint.
Thought I had that interactive thing under control. Well, I guess I did.
But I made a mistake and added the `-a` argument on to `git commit`.
That meant `git` just ignored the staging area, and commited all
changes.

Unfortunately, this means the `test_wikipedia` and
`test_wikipedia_disambiguation` tests got lumped together into one. Not
a huge deal, but I would have preferred them separate for clarity.

The `test_wikipedia_disambiguation` test will query our `WikiProvider`
with a couple words which are known to have disambiguous results.
Because the exception is already handled in `WikiProvider.lookup`, we
are only checking to make sure we recieved the disambiguation error
message.

While we're at it, let's go ahead and `flake8` a couple formatting
issues right on out of here.