Add lookup command
Create a basic lookup command to replace the original stub code. This
defines a similar calling convention compared to the original with some
key differences. The user must now specify the type of provider `lookup`
or `transform`, followed by the full provider name `wikipedia`, for
example. Lastly containing the terms to be looked up, surrounded in
quotes if there are spaces, and with an optional positive integer
`limit` for the number of results to return.

`wt lookup {wikipedia,...} 'search terms' [limit=0]`
`wt lookup wikipedia 'Gentoo Linux'`
`wt lookup urbandictionary hacker 3`

Down the road, I would like to enhance this system to be more user
friendly. Rather than requiring long, specific commands to be entered.
Try to be a little more flexible. Don't require the type of provider,
determine it based on the provider itself. Or create shorter aliases to
specific providers, in an extensible way. Several different
possibilities here worth exploring once the core entry-point is
solidified.