Add transform command
Following the same format as the previous command, make it work for
`TransformProvider`s as well. This is necessary because we are using two
separate collections of providers in the main `word_tools` module --
`.lookup` and `.transform`. They were originally split to give a greater
separation, and clarity between different functions. As well as to
provide more specialized interfaces.

To remedy the code repetition, and lack of flexibility we are seeing
creep back in over here, we may want to rethink how we store our
providers in the main module. Just something to think about.