Re-design CLI arguments
After much back & forth, I am settling in on a CLI argument
specification that I think will work for us. Doing away with the
sub-commands and grouping used originally. Decided to stick everything
into the `main` function. Mostly because that is where everything wound
up when I started to hone in on this rendition.

The new system is not perfect, and I notice some areas I want to improve
on. This includes optional arguments that are specific to different
types of Provider objects. Example of that being the `limit` argument
being specific to `LookupProvider`.

The biggest area in need of improvement is the `try`/`except` blocks
used when creating our concrete Provider objects. I have not determined
a way to grab Providers out of `word_tools` based on whether or not they
are `LookupProvider` or `TransformProvider`. To explain that a bit more
clearly -- we recieve a Provider key on the command line. Depending if
that is a Lookup or Transform, we may call either:

`word_tools.lookup.get(KEY)` or `word_tools.transform.get(KEY)`

I would like to utilize the same call no matter which Provider type we
recieve in the arguments. The same goes for the actual
`provider.lookup` or `provider.convert` calls. If not possible to do
this directly here, look into restructuring the module to accomodate for
that.

The new argument specification brought in here is by no means complete,
or complete. It will change over the course of this branch, and likely a
bit afterwards. The goal is to solidfy the specification and bring it
into use here. Ideally arriving at a system which will expand without
modification as we add new providers in time. For now, the specification
is laid out as :

`$ wt [OPTIONS] PROVIDER WORD`
`$ wt --limit 2 wikipedia "Gentoo Linux"`
`$ wt stoopid "I am very smart"`