Comments: Notes for improvement.
I was considering to further continue tweaking the new CLI
specification. Wrote some comments last night as notes for myself.
Today, I am not sure I'm ready to tackle this bit just yet. Need some
time to think through alternative solutions, and to determine if the one
mentioned in these comments is a valid solution.

I'll get back to this. For now, leaving the comments in as notes behind
the madness here.