Update `test_cli` to use new CLI
Quick update to ensure the `test_display_results` actually calls the
correct function for testing. I think this can be done better. We should
not need to call on `__main__` directly. I think. For now, I just wanted
it to work properly. Going to add this onto the whiteboard for further
review.