Update `README.md`, `CONTRIBUTING.md`
This is the large update to `README.md` that has been a long time
coming. Kinda should have had most of this set up a while back, but we
got distracted and felt the old content was 'good enough'. A comment
from a member of the Python Discord server kindly put it back in my mind
last week that I should get this done. So, here it is. Thank you.

This update modifies most all of the original content in the `README.md`
file to reflect the current state of the project, and expand on the
original information. A summary of these changes includes:

  - Expanded introduction block
  - Add requirements section
  - Clarify installation section, update instructions
  - Explain the thesis written in `CONTRIBUTING.md`
  - Include some words on the project's long-term direction, and goals.

Included with this commit is an update to `CONTRIBUTING.md` as well.
This touches up a few rough edges, and updates a few commands listed to
reflect the current project structure.

This commit resolves issue #14.