Deactivation Periodic Labeller
Over the past couple of weeks, this seems to have been intermittently
(or possibly always?) failing. I am not totally sure why, and have not
felt the pull to look deeper into it as it never seemed to properly
work. In an attempt to put an end to the repeated "Action failed" emails
from GitHub, I am disabling the Action.

Further actions may include looking into the problem and fixing it
properly, so that it actually works like it should. Or just flat out
removing the relevant files and moving on with my life. I have yet to
decide.